+++ Lib/TSCompiler/FastBundleBundler.cs
@@ -128,7 +128,7 @@ namespace Lib.TSCompiler
                 FilesContent[_versionDirPrefix + "loader.js"] = _tools.LoaderJs;
                 if (Project.LiveReloadEnabled)
                 {
-                    FilesContent[_versionDirPrefix + "liveReload.js"] = _tools.LiveReloadJs.Replace("##Idx##", (Project.LiveReloadIdx+1).ToString());
+                    FilesContent[_versionDirPrefix + "liveReload.js"] = _tools.LiveReloadJs.Replace("##Idx##", (Project.LiveReloadIdx + 1).ToString());
                 }
             }
             FilesContent[_versionDirPrefix + PathUtils.WithoutExtension(mapUrl)] = _bundleJs;
@@ -138,7 +138,7 @@ namespace Lib.TSCompiler
 
         void BuildFastBundlerTestHtml(List<string> testSources, string root, string cssLink)
         {
-            var reqSpec = string.Join(' ', testSources.Select(src => "R.r(\'" + PathUtils.WithoutExtension(PathUtils.Subtract(src, root)) + "\');"));
+            var reqSpec = string.Join(' ', testSources.Where(src => !src.EndsWith(".d.ts")).Select(src => "R.r(\'" + PathUtils.WithoutExtension(PathUtils.Subtract(src, root)) + "\');"));
             _indexHtml = $@"<!DOCTYPE html>
 <html>
     <head>
@@ -169,7 +169,7 @@ namespace Lib.TSCompiler
             {
                 liveReloadInclude = $@"<script src=""{_versionDirPrefix}liveReload.js"" charset=""utf-8""></script>";
             }
-                _indexHtml = $@"<!DOCTYPE html>
+            _indexHtml = $@"<!DOCTYPE html>
 <html>
     <head>
         <meta charset=""utf-8"">{Project.HtmlHeadExpanded}